Moved slog logger implementation to providers/slog directory
