+++ providers_test.go
@@ -2,8 +2,6 @@ package vshard_router_test
 
 import (
 	"context"
-	"log/slog"
-	"os"
 	"testing"
 	"time"
 
@@ -49,46 +47,5 @@ func TestStdoutLogger(t *testing.T) {
 	})
 	require.NotPanics(t, func() {
 		stdoutLogger.Debugf(ctx, "")
-	})
-}
-
-func TestNewSlogLogger(t *testing.T) {
-	var slogProvider vshardrouter.LogfProvider
-
-	require.NotPanics(t, func() {
-		slogProvider = vshardrouter.NewSlogLogger(nil)
-	})
-
-	require.Panics(t, func() {
-		slogProvider.Warnf(context.TODO(), "")
-	})
-}
-
-func TestSlogProvider(t *testing.T) {
-	ctx := context.Background()
-
-	// create new logger handler
-	handler := slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{
-		Level: slog.LevelError,
-	})
-	// create new SLogger instance
-	sLogger := slog.New(handler)
-
-	logProvider := vshardrouter.NewSlogLogger(sLogger)
-
-	require.NotPanics(t, func() {
-		logProvider.Infof(ctx, "test %s", "s")
-	})
-
-	require.NotPanics(t, func() {
-		logProvider.Warnf(ctx, "test %s", "s")
-	})
-
-	require.NotPanics(t, func() {
-		logProvider.Errorf(ctx, "test %s", "s")
-	})
-
-	require.NotPanics(t, func() {
-		logProvider.Debugf(ctx, "test %s", "s")
 	})
 }
